% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echogo_html_helpers.R
\name{.echogo_html_details}
\alias{.echogo_html_details}
\title{Create an HTML details wrapper}
\usage{
.echogo_html_details(summary, inner_html)
}
\description{
Create an HTML details wrapper
}
\keyword{internal}
