% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_render.R
\name{.render_echogo_report}
\alias{.render_echogo_report}
\title{Render the EchoGO HTML report}
\usage{
.render_echogo_report(
  report_title,
  template = NULL,
  params = list(),
  theme = "flatly",
  outdir,
  keep_temp_rmd = FALSE,
  verbose_render = FALSE
)
}
\description{
Render the EchoGO HTML report
}
\keyword{internal}
