% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species.R
\name{echogo_add_tags}
\alias{echogo_add_tags}
\title{Add curated tag panels from inst/extdata/species_tags.yaml}
\usage{
echogo_add_tags(
  df,
  yaml_path = system.file("extdata", "species_tags.yaml", package = "EchoGO")
)
}
\arguments{
\item{df}{Species data frame containing an `organism` column.}

\item{yaml_path}{Path to a species tag YAML file.}
}
\description{
Add curated tag panels from inst/extdata/species_tags.yaml
}
