% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_helpers.R
\name{echogo_demo_path}
\alias{echogo_demo_path}
\title{Path to shipped demo inputs}
\usage{
echogo_demo_path()
}
\value{
A filesystem path to inst/extdata/echogo_demo
}
\description{
Path to shipped demo inputs
}
