% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgdb_helpers.R
\name{echogo_list_orgdb}
\alias{echogo_list_orgdb}
\title{List available OrgDb annotation packages}
\usage{
echogo_list_orgdb(
  pkgs = c("org.Hs.eg.db", "org.Mm.eg.db", "org.Dr.eg.db", "org.Dm.eg.db",
    "org.Rn.eg.db", "org.Ce.eg.db")
)
}
\arguments{
\item{pkgs}{Character vector of OrgDb package names to check.
Defaults to a common set: human, mouse, zebrafish, fly, rat, worm.}
}
\value{
A data.frame with columns: package, installed (TRUE/FALSE), and loadable (TRUE/FALSE).
}
\description{
Shows which organism annotation databases (OrgDb) are installed and which are missing.
}
\examples{
echogo_list_orgdb()
echogo_list_orgdb(c("org.Hs.eg.db", "org.Mm.eg.db"))
}
