% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species.R
\name{echogo_pick_species}
\alias{echogo_pick_species}
\title{Pick species interactively (always shows a view)}
\usage{
echogo_pick_species(refresh = FALSE, page_len = 25, force_browser = FALSE)
}
\arguments{
\item{refresh}{Logical; refresh species metadata before displaying it.}

\item{page_len}{Number of rows per page in the interactive table.}

\item{force_browser}{Logical; save and open a standalone widget in a browser.}
}
\description{
Pick species interactively (always shows a view)
}
