% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_helpers.R
\name{echogo_scaffold}
\alias{echogo_scaffold}
\title{Scaffold a minimal EchoGO project directory}
\usage{
echogo_scaffold(path = "echogo_project")
}
\arguments{
\item{path}{target directory}
}
\description{
Scaffold a minimal EchoGO project directory
}
