% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species.R
\name{echogo_validate_species}
\alias{echogo_validate_species}
\title{Validate and/or suggest species IDs for `run_full_echogo()`}
\usage{
echogo_validate_species(species, refresh = FALSE)
}
\arguments{
\item{species}{Character vector of g:Profiler organism IDs or names.}

\item{refresh}{Logical; refresh the species cache before validation.}
}
\description{
Validate and/or suggest species IDs for `run_full_echogo()`
}
