% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_visuals.R
\name{plot_echogo_context_summary}
\alias{plot_echogo_context_summary}
\title{Plot how selected annotation contexts represent the same experimental query}
\usage{
plot_echogo_context_summary(sources, evidence, contexts)
}
\description{
Plot how selected annotation contexts represent the same experimental query
}
\keyword{internal}
