% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_visuals.R
\name{plot_echogo_profile_counts}
\alias{plot_echogo_profile_counts}
\title{Plot primary EchoGO evidence-profile counts}
\usage{
plot_echogo_profile_counts(evidence)
}
\description{
Plot primary EchoGO evidence-profile counts
}
\keyword{internal}
