% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_visuals.R
\name{plot_echogo_representability}
\alias{plot_echogo_representability}
\title{Plot representability from the target experiment to comparative submission}
\usage{
plot_echogo_representability(mapping)
}
\description{
Plot representability from the target experiment to comparative submission
}
\keyword{internal}
