% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_eval.R
\name{summarize_eqi_distributions}
\alias{summarize_eqi_distributions}
\title{Summarize EQI distributions and enrichment strength (PDF + Excel)
Mirrors Rmd "EQI_distribution_by_ontology.pdf" and FE distribution}
\usage{
summarize_eqi_distributions(consensus_all, goseq, output_dir, mode = "with_bg")
}
\description{
Summarize EQI distributions and enrichment strength (PDF + Excel)
Mirrors Rmd "EQI_distribution_by_ontology.pdf" and FE distribution
}
\keyword{internal}
