% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species.R
\name{echogo_find_species}
\alias{echogo_find_species}
\title{Search EchoGO's species list by name/ID/alias/NCBI}
\usage{
echogo_find_species(
  query,
  fields = c("organism", "name", "alias", "ncbi"),
  fuzzy = TRUE,
  refresh = FALSE
)
}
\description{
Search EchoGO's species list by name/ID/alias/NCBI
}
