% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species.R
\name{echogo_pick_species}
\alias{echogo_pick_species}
\title{Pick species interactively (always shows a view)}
\usage{
echogo_pick_species(refresh = FALSE, page_len = 25, force_browser = FALSE)
}
\description{
Pick species interactively (always shows a view)
}
