% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species.R
\name{echogo_preflight_species}
\alias{echogo_preflight_species}
\title{Check species argument and provide actionable feedback (preflight)}
\usage{
echogo_preflight_species(species, refresh = FALSE, error_if_unknown = TRUE)
}
\description{
Check species argument and provide actionable feedback (preflight)
}
