% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species.R
\name{echogo_species_lookup}
\alias{echogo_species_lookup}
\title{Suggest best-matching g:Profiler IDs for each input (smart ranking)}
\usage{
echogo_species_lookup(
  query,
  fields = c("organism", "name", "alias", "ncbi"),
  top_n = 5,
  fuzzy = TRUE,
  refresh = FALSE
)
}
\description{
Suggest best-matching g:Profiler IDs for each input (smart ranking)
}
