% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species.R
\name{echogo_update_species_cache}
\alias{echogo_update_species_cache}
\title{Refresh and persist the g:Profiler species list}
\usage{
echogo_update_species_cache(update_package_data = FALSE, verbose = TRUE)
}
\value{
tibble with columns: organism, name, ncbi, alias
}
\description{
Refresh and persist the g:Profiler species list
}
