% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{score_consensus_terms}
\alias{score_consensus_terms}
\title{Score and classify consensus enrichment terms}
\usage{
score_consensus_terms(
  df,
  P_CAP = 6,
  FOLD_CAP = 8,
  W_STRICT = list(w_goseq = 1, w_bgprev = 0.8, w_foldbg = 0.4, w_foldgs = 0.4, w_p = 0.4),
  W_ALL = list(w_goseq = 1, w_bgprev = 0.7, w_nbprev = 0.3, w_foldbg = 0.35, w_foldnb =
    0.25, w_foldgs = 0.35, w_p = 0.35)
)
}
\arguments{
\item{df}{Data frame output from `merge_enrichment_sources()`.}
}
\value{
Scored consensus enrichment table.
}
\description{
Applies metrics such as species support, consensus scores, robustness, origin tags, etc.
}
\keyword{internal}
